+++ app/examples/column-resizing/page.tsx
@@ -44,7 +44,7 @@ export default async function DraggingExamplePage() {
           add a <Code>resizable: true</Code> property to each <strong>resizing candidate</strong> column;
         </li>
         <li>
-          use <Code>useDragToggleColumns()</Code> hook to get the effective columns.
+          use the <Code>useDataTableColumns()</Code> hook to get the effective columns.
         </li>
       </UnorderedList>
       <CodeBlock code={code['ResizingExample.tsx']} />